fix(protocol): change transition ID to uint24 and fix a liveness bond return bug (#17830)

+++ packages/protocol/contracts/L1/libs/LibVerifying.sol
@@ -16,13 +16,13 @@ library LibVerifying {
         uint64 blockId;
         uint64 slot;
         uint64 numBlocksVerified;
-        uint32 tid;
-        uint32 lastVerifiedTransitionId;
+        uint24 tid;
+        uint24 lastVerifiedTransitionId;
         uint16 tier;
         bytes32 blockHash;
         bytes32 syncStateRoot;
         uint64 syncBlockId;
-        uint32 syncTransitionId;
+        uint24 syncTransitionId;
         address prover;
         bool postFork;
         ITierRouter tierRouter;
